refactor(dot11): split layer and element tests into tests layer module
